Eric Sunshine eliminated compilation warning in terrain2/bruteblockrenderer.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@29233 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/terrain/terrain2/bruteblockrenderer.cpp
@@ -929,7 +929,7 @@ void TerrainBlock::CullRenderMeshes (iRenderView* rview, const csPlane3* cullPla
     meshCache.Push (mesh);
   }
 
-  for (int j = 0; j < renderData->alphaMapArrayAlpha.GetSize (); ++j)
+  for (size_t j = 0; j < renderData->alphaMapArrayAlpha.GetSize (); ++j)
   {
     iMaterialWrapper* mat = renderData->materialArrayAlpha[j];
     iShaderVariableContext* svContext = renderData->svContextArrayAlpha[j];
